Change local secret manager by noop
