+++ openmetadata-core/src/main/resources/json/schema/entity/services/connections/metadata/secretsManagerProvider.json
@@ -5,6 +5,6 @@
   "description": "OpenMetadata Secrets Manager Provider. Make sure to configure the same secrets manager providers as the ones configured on the OpenMetadata server.",
   "type": "string",
   "javaType": "org.openmetadata.catalog.services.connections.metadata.SecretsManagerProvider",
-  "enum": ["local", "aws", "aws-ssm"],
+  "enum": ["noop", "aws", "aws-ssm"],
   "additionalProperties": false
 }
